requirements.txt
flask
gunicorn
Pillow
numpy
matplotlib
